chore: move preview card generation into own file
