+++ gatsby/util/preview-card-generator/assets/coco-bold.ttf
